게시판 페이지 목록 화면 구성

+++ board/src/main/java/com/example/demo/entity/Board.java
@@ -16,7 +16,7 @@ public class Board {
     // PK
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     // 작성자 이름
     @Column(length = 50)
@@ -37,7 +37,7 @@ public class Board {
     private LocalDateTime updateTime;
 
     @Builder
-    public Board(int id, String userName, String title, String contents, LocalDateTime createTime, LocalDateTime updateTime) {
+    public Board(Long id, String userName, String title, String contents, LocalDateTime createTime, LocalDateTime updateTime) {
         this.id = id;
         this.userName = userName;
         this.title = title;
